money_agent/requirements.txt
# Core Dependencies
anthropic>=0.40.0
notion-client>=2.2.0
discord.py>=2.3.0
pydantic>=2.0.0

# Scheduling
apscheduler>=3.10.0
pytz>=2024.1

# Utilities
python-dotenv>=1.0.0
rich>=13.7.0

# Date handling
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
